pyproject.toml
[project]
name = "meetingassistant"
version = "0.1.0"
description = ""
authors = [
    { name = "fwang02", email = "fywang021205@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "openai (>=1.72.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pylint (>=3.3.6,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "openai-whisper (>=20240930,<20240931)",
    "poetry-core (>=2.0.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"